change port angle type annotation from int to float
Former-commit-id: 1a63d6348ca7bea38bdba6d4bebea82474f3b434 [formerly b397f03b8c36a341a0bfb7ad8e322a9d4716d1aa [formerly 710fc1ba7f6910ef85f44e14a1c0ccea625e98e0]]
Former-commit-id: b397f03b8c36a341a0bfb7ad8e322a9d4716d1aa
Former-commit-id: 4106c6983e79cbf298a648cfc6336cf2fd0671cc

+++ gdsfactory/components/bend_circular_heater.py
@@ -10,7 +10,7 @@ from gdsfactory.types import CrossSectionFactory
 @gf.cell
 def bend_circular_heater(
     radius: float = 10,
-    angle: int = 90,
+    angle: float = 90,
     npoints: int = 720,
     heater_to_wg_distance: float = 1.2,
     heater_width: float = 0.5,
